+++ src/Format/Reader/SimpleReader.php
@@ -22,7 +22,6 @@ declare(strict_types=1);
 
 namespace FormatPHP\Format\Reader;
 
-use FormatPHP\ConfigInterface;
 use FormatPHP\Exception\InvalidMessageShapeException;
 use FormatPHP\Format\ReaderInterface;
 use FormatPHP\Format\Writer\SimpleWriter;
@@ -45,7 +44,7 @@ class SimpleReader implements ReaderInterface
     /**
      * @inheritdoc
      */
-    public function __invoke(ConfigInterface $config, array $data): MessageCollection
+    public function __invoke(array $data): MessageCollection
     {
         $messages = new MessageCollection();