+++ tests/Util/MockFormatReader.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace FormatPHP\Test\Util;
 
-use FormatPHP\ConfigInterface;
 use FormatPHP\Format\ReaderInterface;
 use FormatPHP\MessageCollection;
 
@@ -13,7 +12,7 @@ class MockFormatReader implements ReaderInterface
     /**
      * @inheritdoc
      */
-    public function __invoke(ConfigInterface $config, array $data): MessageCollection
+    public function __invoke(array $data): MessageCollection
     {
         return new MessageCollection();
     }