+++ tests/Util/fixtures/reader-closure-invalid-05.php
@@ -2,10 +2,9 @@
 
 declare(strict_types=1);
 
-use FormatPHP\ConfigInterface;
 use FormatPHP\MessageCollection;
 
-return function (ConfigInterface $config, array $data)
+return function (array $data)
 {
     return new MessageCollection();
 };